Move allocationToParams, create guaranteeToParams
